# repo/api/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CineRide Media Operations Platform API",
    "version": "1.0.0",
    "description": "REST API for the CineRide offline-first media operations platform. Manages movies, editorial content, ride requests, environmental sensors, and a local funds ledger."
  },
  "servers": [
    { "url": "/api", "description": "Local API (via Nginx reverse proxy)" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
  },
  "paths": {
...
    "/movies/{id}": {
      "get": { "summary": "Get movie by ID", "tags": ["Movies"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Movie detail" } } },
      "put": { "summary": "Update movie (staff only)", "tags": ["Movies"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Movie updated" } } },
      "delete": { "summary": "Delete movie (staff only)", "tags": ["Movies"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Movie deleted" } } }
    },
...
    "/movie-import/upload": {
      "post": { "summary": "Upload import file (JSON/CSV)", "tags": ["Movie Import"], "security": [{ "BearerAuth": [] }], "responses": { "201": { "description": "Import job created" } } }
    },
    "/movie-import/{jobId}": {
      "get": { "summary": "Get import job status", "tags": ["Movie Import"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Import job with parsed records and conflicts" } } }
    },
    "/movie-import/{jobId}/resolve/{recordIndex}": {
      "put": { "summary": "Resolve conflicts for a record", "tags": ["Movie Import"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }, { "name": "recordIndex", "in": "path", "required": true, "schema": { "type": "integer" } }], "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "resolutions": { "type": "object", "description": "Map of conflict field to resolution value" } } } } } }, "responses": { "200": { "description": "Conflict resolved" } } }
    },
    "/movie-import/{jobId}/skip/{recordIndex}": {
      "post": { "summary": "Skip a conflicted record", "tags": ["Movie Import"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }, { "name": "recordIndex", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Record skipped" } } }
    },
    "/movie-import/{jobId}/execute": {
      "post": { "summary": "Execute import after conflict resolution", "tags": ["Movie Import"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Import executed" } } }
    },
...
    "/search": {
      "get": { "summary": "Unified search across movies, content, users", "tags": ["Search"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "q", "in": "query", "required": true, "schema": { "type": "string" } }, { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["movie", "content", "user"] } }, { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["popularity", "rating", "newest"] } }], "responses": { "200": { "description": "Search results" } } }
    },
    "/search/suggest": {
      "get": { "summary": "Get search suggestions (typo-tolerant)", "tags": ["Search"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "q", "in": "query", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Suggestions" } } }
    },
...
    "/sensors/ingest": {
      "post": { "summary": "Ingest a single sensor reading", "tags": ["Sensors"], "description": "Requires X-Device-Id and X-Device-Secret headers. Body device_id must match authenticated device.", "parameters": [{ "name": "X-Device-Id", "in": "header", "required": true, "schema": { "type": "string" } }, { "name": "X-Device-Secret", "in": "header", "required": true, "schema": { "type": "string" } }], "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "device_id": { "type": "string" }, "timestamp": { "type": "string", "format": "date-time" }, "value": { "type": "number" } } } } } }, "responses": { "201": { "description": "Reading ingested" }, "401": { "description": "Invalid device credentials" }, "403": { "description": "Body/header device_id mismatch" } } }
    },
    "/sensors/ingest/batch": {
      "post": { "summary": "Batch ingest sensor readings (resumable)", "tags": ["Sensors"], "description": "Requires X-Device-Id and X-Device-Secret headers. All readings must match authenticated device. Optional X-Batch-Session header for resumable transfer.", "parameters": [{ "name": "X-Device-Id", "in": "header", "required": true, "schema": { "type": "string" } }, { "name": "X-Device-Secret", "in": "header", "required": true, "schema": { "type": "string" } }, { "name": "X-Batch-Session", "in": "header", "schema": { "type": "string" } }], "responses": { "201": { "description": "Batch results with session_id for resume" }, "401": { "description": "Invalid device credentials" }, "403": { "description": "Batch contains mismatched device_id" } } }
    },
    "/sensors/devices": {
      "get": { "summary": "List sensor devices", "tags": ["Sensors"], "security": [{ "BearerAuth": [] }], "responses": { "200": { "description": "Device list" } } },
      "post": { "summary": "Register sensor device (admin only)", "tags": ["Sensors"], "security": [{ "BearerAuth": [] }], "responses": { "201": { "description": "Device created" } } }
    },
...
    "/sensors/devices/{id}": {
      "get": { "summary": "Get sensor device detail", "tags": ["Sensors"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Device detail" } } },
      "put": { "summary": "Update sensor device (admin)", "tags": ["Sensors"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Device updated" } } }
    },
    "/sensors/devices/{id}/rotate-secret": {
      "post": { "summary": "Rotate device secret (admin)", "tags": ["Sensors"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "New secret returned" } } }
    },
    "/sensors/readings/{deviceId}": {
      "get": { "summary": "Get sensor readings", "tags": ["Sensors"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "deviceId", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Reading list" } } }
    },
...
  }
}